Allow profile IDs up to 255 (NO_PROFILE stays 200)
NUM_PROFILES and NUM_RECRUITABLE are 256; IDs 165-199 and 201-254 are for
added characters (200 is NO_PROFILE, 255 is a 'none' value in a few
places). The IDs of the friends and enemies of a profile are INT16 now
(prof.dat is still read as bytes), so they are no longer limited to 127.
Loops over the profile IDs use size_t, the schemas and validations no
longer stop at 199, MAX_NUMBER_MERCS is 130 and SAVE_GAME_VERSION is 107.

+++ src/game/Laptop/AIMMembers.cc
@@ -1802,7 +1802,7 @@ static BOOLEAN CanMercBeHired(void)
 	for (UINT8 i = HATED_SLOT1; i < NUM_HATED_SLOTS; ++i)
 	{
 		//see if someone the merc hates is on the team
-		INT8 const bMercID = p.bHated[i];
+		INT16 const bMercID = p.bHated[i];
 		if (bMercID < 0) continue;
 
 		if (!IsMercOnTeamAndInOmertaAlreadyAndAlive(bMercID)) continue;
